wcc dist queue preprocess implementation

+++ includes/mincut_custom.h
@@ -9,16 +9,16 @@ class MinCutCustom {
     public:
         MinCutCustom(const igraph_t* graph) : graph(graph) {
         };
-        int ComputeMinCut();
+        long ComputeMinCut();
         /* int ComputeAllMinCuts(); */
-        const std::vector<int>& GetInPartition() const;
-        const std::vector<int>& GetOutPartition() const;
+        const std::vector<long>& GetInPartition() const;
+        const std::vector<long>& GetOutPartition() const;
         /* const std::vector<std::vector<int>>& GetAllPartitions() const; */
     private:
-        std::vector<int> in_partition;
-        std::vector<int> out_partition;
+        std::vector<long> in_partition;
+        std::vector<long> out_partition;
         /* std::vector<std::vector<int>> all_partitions; */
-        std::map<int, int> new_to_old_node_id_map;
+        std::map<long, long> new_to_old_node_id_map;
         const igraph_t* graph;
 };
